Create new `h.schemas.forms.accounts.edit_profile` module and tests
Update `h.views.accounts` to use new `EditProfileSchema`
Remove `EditProfileSchema` from `h.accounts.schemas`
