+++ h/schemas/forms/accounts/__init__.py
@@ -1,3 +1,10 @@
 # -*- coding: utf-8 -*-
 
 from __future__ import unicode_literals
+
+from h.schemas.forms.accounts.edit_profile import EditProfileSchema
+
+
+__all__ = (
+    'EditProfileSchema',
+)
